test(api): mirror-path test files for composio *_tools schemas + fix mutmut classmethod lane
PR #922's mutation gate flags app/models/composio_schemas/{github,linear,slack}_tools.py as changed modules with no mapped test file. The shared test_composio_schemas.py imports from the package __init__, so the AST importer never maps it to the *_tools modules.

- split the *_tools coverage out of test_composio_schemas.py into mirror-path files (tests/unit/models/composio_schemas/test_{github,linear,slack}_tools.py) that import the modules directly; 154 tests, all green
- root-fix mutmut_decorated_patch.py: it stripped @classmethod from the mangled orig/mutant copies, leaving plain functions in the class dict — the trampoline's getattr(cls, ...) returned them UNBOUND and every classmethod call died with 'missing 1 required positional argument'. Keep @classmethod/@staticmethod on the copies (harmless to re-execute, required for the binding dance); strip everything else as before.

Lanes verified locally: github_tools 11/11 killed, linear_tools + slack_tools zero survivors.
